move queries to the query directory, keep deserializer in repository
