+++ backend/infrahub/core/diff/query/delete_query.py
