+++ backend/infrahub/core/diff/query/get_conflict_query.py
@@ -1,19 +1,17 @@
 from typing import Any
 
-from infrahub.core.diff.model.path import EnrichedDiffConflict
+from neo4j.graph import Node as Neo4jNode
+
 from infrahub.core.query import Query, QueryType
 from infrahub.database import InfrahubDatabase
-
-from .deserializer import EnrichedDiffDeserializer
 
 
 class EnrichedDiffConflictQuery(Query):
     name = "enriched_diff_conflict"
     type = QueryType.READ
 
-    def __init__(self, deserializer: EnrichedDiffDeserializer, conflict_id: str, **kwargs: Any) -> None:
+    def __init__(self, conflict_id: str, **kwargs: Any) -> None:
         super().__init__(**kwargs)
-        self.deserializer = deserializer
         self.conflict_id = conflict_id
 
     async def query_init(self, db: InfrahubDatabase, **kwargs: Any) -> None:
@@ -22,8 +20,8 @@ class EnrichedDiffConflictQuery(Query):
         self.return_labels = ["conflict"]
         self.add_to_query(query=query)
 
-    async def get_conflict(self) -> EnrichedDiffConflict | None:
+    async def get_conflict_node(self) -> Neo4jNode | None:
         result = self.get_result()
         if not result:
             return None
-        return self.deserializer.deserialize_conflict(diff_conflict_node=result.get_node("conflict"))
+        return result.get_node("conflict")
