+++ backend/infrahub/core/diff/query/update_conflict_query.py
@@ -1,10 +1,10 @@
 from typing import Any
 
-from infrahub.core.diff.model.path import ConflictSelection, EnrichedDiffConflict
+from neo4j.graph import Node as Neo4jNode
+
+from infrahub.core.diff.model.path import ConflictSelection
 from infrahub.core.query import Query, QueryType
 from infrahub.database import InfrahubDatabase
-
-from .deserializer import EnrichedDiffDeserializer
 
 
 class EnrichedDiffConflictUpdateQuery(Query):
@@ -13,13 +13,11 @@ class EnrichedDiffConflictUpdateQuery(Query):
 
     def __init__(
         self,
-        deserializer: EnrichedDiffDeserializer,
         conflict_id: str,
         selection: ConflictSelection | None,
         **kwargs: Any,
     ) -> None:
         super().__init__(**kwargs)
-        self.deserializer = deserializer
         self.conflict_id = conflict_id
         self.selection = selection
 
@@ -32,8 +30,8 @@ class EnrichedDiffConflictUpdateQuery(Query):
         self.return_labels = ["conflict"]
         self.add_to_query(query=query)
 
-    async def get_conflict(self) -> EnrichedDiffConflict | None:
+    async def get_conflict_node(self) -> Neo4jNode | None:
         result = self.get_result()
         if not result:
             return None
-        return self.deserializer.deserialize_conflict(diff_conflict_node=result.get_node("conflict"))
+        return result.get_node("conflict")
